requirements.txt
feedparser>=6.0.10
Pillow>=10.0.0
moviepy>=1.0.3
schedule>=1.2.0
pyyaml>=6.0.1
dashscope>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0 
markdown2>=2.4.0